simple stack is done
