--- src/homework_4/gevorg_arghushyan/chapter_6/sipleStack/StackDemo.java
@@ -1,22 +1,0 @@
-package homework_4.gevorg_arghushyan.chapter_6.sipleStack;
-
-public class StackDemo {
-    public static void main(String[] args) {
-        Stack myStakc = new Stack();
-        myStakc.push(1);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-        myStakc.push(2);
-    }
-}
