+++ src/homework_4/gevorg_arghushyan/sipleStack/StackDemo.java
@@ -1,0 +1,16 @@
+package homework_4.gevorg_arghushyan.sipleStack;
+
+public class StackDemo {
+    public static void main(String[] args) {
+        Stack myStack = new Stack(4);
+        myStack.push(2);
+        myStack.push(2);
+        myStack.push(2);
+        myStack.push(2);
+        myStack.push(2);
+        myStack.push(2);
+        myStack.push(2);
+        myStack.push(2);
+
+    }
+}
